+++ module/items/sheets/shot.mjs
@@ -1,0 +1,39 @@
+import {OutgunnedUtilities} from "../../apps/utilities.mjs";
+
+export class OutgunnedShotSheet extends ItemSheet {
+
+
+  static get defaultOptions() {
+    return foundry.utils.mergeObject(super.defaultOptions, {
+      classes: ["outgunned", "sheet", "item"],
+      width: 520,
+      height: 600,
+      tabs: [{ navSelector: ".sheet-tabs", contentSelector: ".sheet-body", initial: "description" }]
+    });
+  }
+
+  get template() {
+    const path = "systems/outgunned/templates/item";
+    return `${path}/${this.item.type}.html`;
+  }
+
+  getData() {
+    const context = super.getData();
+    const itemData = context.item;
+    context.rollData = {};
+    let actor = this.object?.parent ?? null;
+    if (actor) {
+      context.rollData = actor.getRollData();
+    }
+    context.system = itemData.system;
+    context.flags = itemData.flags;
+    context.isGM =  game.user.isGM;
+    context.itemType = game.i18n.localize('OG.'+itemData.system.type)
+    return context;
+  }
+
+  activateListeners(html) {
+    super.activateListeners(html);
+    if (!this.isEditable) return;
+  }
+}